refactor: background 파일 분리=
- manifest에서 background scripts 를 page로 바꾼 후 background/index.html로 변경
- background/index.html 생성 후 background/index.js 모듈 방식으로 import
- background.js 파일 분리
