+++ js/background/index.html
@@ -1,0 +1,1 @@
+<script type="module" src="./index.js"></script>
